JAMES-2529 Extract ContentMatcher and HeaderExtractor from MailMatcher
